Enable CppInterOp
[CppInterOp](https://github.com/compiler-research/CppInterOp) exposes API from
Clang and LLVM in a mostly backward compatibe way. The API support downstream
tools that utilize interactive C++ by using the compiler as a service.

Adopting more of CppInterOp in ROOT will help simplify the LLVM migration
process and allow us to upstream more patches to either CppInterOp or LLVM.